requirements.txt
# Azure Enterprise Data Platform Requirements
# Alex Taylor Finch Cognitive Architecture - Version 1.0.0 UNNILNILIUM
# Python Virtual Environment Dependencies
# Generated: August 7, 2025
# Updated: August 7, 2025 - Added OneLake data analysis dependencies
# Updated: August 8, 2025 - Added cpestaginglake discovery notebook dependencies

# Environment Configuration
python-dotenv>=1.0.0

# Core Azure SDK and Authentication
azure-core>=1.29.0
azure-identity>=1.14.0
azure-mgmt-core>=1.4.0

# Azure Service Management
azure-mgmt-resource>=23.0.0
azure-mgmt-sql>=3.0.0
azure-mgmt-synapse>=2.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-datafactory>=4.0.0

# Azure Data Services
azure-storage-blob>=12.17.0
azure-storage-file-datalake>=12.12.0
azure-synapse-artifacts>=0.17.0
azure-synapse-spark>=0.7.0

# Microsoft Fabric and OneLake Support
# Note: azure-storage-file 2.1.0+ conflicts with azure-cli
# Use azure-storage-file-datalake for modern OneLake/ADLS Gen2 support
azure-storage-file-datalake>=12.12.0
azure-data-tables>=12.4.0

# Apache Spark ecosystem
pyspark>=3.4.0
py4j>=0.10.9.7
delta-spark>=2.4.0
koalas>=0.32.0

# PySpark SQL and Functions
pyspark[sql]>=3.4.0
# Note: pyspark.sql.functions (col, when, etc.) included in main pyspark package

# Fabric-specific Libraries
# Note: These are typically pre-installed in Microsoft Fabric environment
# For local development, PySpark requires Java 8+ installation
# In Microsoft Fabric, Spark is pre-configured and ready to use
# notebookutils - Available only in Microsoft Fabric environment
# mssparkutils - Available only in Microsoft Fabric environment
# pyfabric>=1.0.0  # Uncomment when available
# fabric-spark>=1.0.0  # Uncomment when available

# Database Connectivity
pyodbc>=4.0.39
sqlalchemy>=2.0.0
pymssql>=2.2.8

# Data Processing and Analytics
pandas>=2.0.0
numpy>=1.25.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Additional Data Analysis Libraries
statsmodels>=0.14.0
pyarrow>=13.0.0
fastparquet>=2023.7.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter and Notebook Support
jupyter>=1.0.0
notebook>=7.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Scientific Computing and Machine Learning
scipy>=1.11.0
scikit-learn>=1.3.0

# Enhanced Machine Learning Libraries
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Advanced Analytics and Statistics
pingouin>=0.5.3
feature-engine>=1.6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Enhanced Visualization Libraries
bokeh>=3.2.0
altair>=5.0.0
wordcloud>=1.9.0

# Big data and distributed computing
dask>=2023.7.0
modin[all]>=0.21.0
polars>=0.18.0

# Memory and Performance Optimization
memory-profiler>=0.61.0
psutil>=5.9.0

# Configuration and Environment Management
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.2

# Logging and Monitoring
structlog>=23.1.0
rich>=13.5.0

# HTTP and API Communication
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# Async Support
asyncio-mqtt>=0.13.0
aiofiles>=23.2.0

# Utilities
click>=8.1.0
typer>=0.9.0
python-dateutil>=2.8.0

# Security and Cryptography
cryptography>=41.0.0
keyring>=24.2.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Enterprise Integration (Optional - Install if needed)
# azure-servicebus>=7.11.0
# azure-eventhub>=5.11.0
# azure-functions>=1.18.0
# azure-cosmosdb-table>=1.0.6

# CPE Staging Lake Discovery Dependencies
# Required for cpestaginglake-discovery-mapping.ipynb
# These are automatically available in Microsoft Fabric environment:
# - notebookutils.mssparkutils (Fabric-native file system operations)
# - spark session (Pre-configured Spark context)
# - PySpark SQL functions (col, when, desc, etc.)
#
# For local development simulation (limited functionality):
# pyspark[sql]>=3.4.0  # Already included above
# pandas>=2.0.0        # Already included above
# Note: Full Fabric functionality requires Microsoft Fabric environment

# Development Dependencies (Optional - Install for development)
# pre-commit>=3.3.0
# bandit>=1.7.0
# safety>=2.3.0
# interrogate>=1.5.0
